2-function-with-parameters.py
Function with Parameters:
The function greet_person(name) accepts a parameter name. This parameter allows the function to use the value passed when the function is called to customize the greeting message.

Defining the Function:
def greet_person(name): defines the function. name is the parameter that holds the value passed when the function is called.

Calling the Function:
greet_person("Wigor") calls the function, passing "Wigor" as the argument. The function then prints "Hi Wigor!".