2-function-with-parameters.py
Function with Parameters:
The function greet_person(name) accepts a parameter name. This parameter allows the function to use the passed value to customize the greeting message.

Defining the Function:
def greet_person(name): defines the function, and name is the parameter that will hold the value passed when the function is called.

Calling the Function:
greet_person("wigor") calls the function, passing the string "wigor" as the argument. The function uses this value to print "Hi wigor!".