5-function-inside-program-flow.py
Function Inside a Program Flow:
The function check_even(number) is integrated into the program's flow, where it performs a task (checking if a number is even or odd) as part of the overall logic of the program.

Defining the Function:
def check_even(number): defines the function check_even(), which takes one parameter, number. Inside the function, an if-else condition is used to determine if the number is even or odd.

Calling the Function:
The function is called twice with different numbers:

check_even(10) checks if 10 is even or odd.

check_even(7) checks if 7 is even or odd.