8-function-with-default-parameter-value.py
he function greet(name="Guest") is defined to take one parameter:

name: Represents the name to greet. If no name is provided, it defaults to "Guest".

Inside the function:

The function prints a greeting message: "Welcome," followed by the value of the name parameter.

If a name is provided, the greeting will use that name.

If no name is provided, the default value "Guest" is used.